% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChain.r
\docType{methods}
\name{absorbingStates}
\alias{absorbingStates}
\alias{absorbingStates,MarkovChain-method}
\title{Returns All Absorbing States}
\usage{
absorbingStates(object)
}
\arguments{
\item{object}{An instance of the \code{MarkovChain}-class}
}
\description{
Returns All Absorbing States
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"MarkovChain\")")}{ Returns the names of all states that never have a successor
in a clickstream (i.e. that are absorbing).} }
}

\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}
