\name{absorbingStates-methods}
\docType{methods}
\alias{absorbingStates}
\alias{absorbingStates,MarkovChain-method}
\title{Returns all absorbing states}
\description{
Returns the names of all states that never have a successor in a clickstream (i.e. that are absorbing).
}
\section{Methods}{
\describe{

\item{\code{signature(object = "MarkovChain")}}{
%%  ~~describe this method here~~
}
}}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}
