% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.r
\docType{class}
\name{EvaluationResult-class}
\alias{EvaluationResult-class}
\title{Class \code{EvaluationResult}}
\description{
Class \code{EvaluationResult}
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("EvaluationResult", ...)}. This S4 class describes \code{EvaluationResult} objects.
}

\examples{

# show EvaluationResult definition
showClass("EvaluationResult")

}
\seealso{
\code{\link{mcEvaluate}}
}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{classes}
