% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.r
\docType{methods}
\name{show,EvaluationResult-method}
\alias{show,EvaluationResult-method}
\title{Shows an \code{EvaluationResult} object}
\usage{
\S4method{show}{EvaluationResult}(object)
}
\arguments{
\item{object}{An instance of the \code{EvaluationResult}-class}
}
\description{
Shows an \code{EvaluationResult} object
}
\section{Methods}{
 \describe{
\item{list("signature(object = \"EvaluationResult\")")}{ Shows an \code{EvaluationResult} object. } }
}

\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}
