% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChain.r
\docType{methods}
\name{show,MarkovChain-method}
\alias{show,MarkovChain-method}
\title{Shows a \code{MarkovChain} object}
\usage{
\S4method{show}{MarkovChain}(object)
}
\arguments{
\item{object}{An instance of the \code{MarkovChain}-class}
}
\description{
Shows a \code{MarkovChain} object
}
\section{Methods}{
 \describe{
\item{list("signature(object = \"MarkovChain\")")}{ Shows an \code{MarkovChain} object. } }
}

\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}
