% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChain.r
\docType{methods}
\name{plot,MarkovChain-method}
\alias{plot,MarkovChain-method}
\title{Plots a \code{MarkovChain} object}
\usage{
\S4method{plot}{MarkovChain}(x, order = 1, digits = 2, minProbability = 0, ...)
}
\arguments{
\item{x}{An instance of the \code{MarkovChain}-class}

\item{order}{The order of the transition matrix that should be plotted}

\item{digits}{The number of digits of the transition probabilities}

\item{minProbability}{Only transitions with a probability >= the specified minProbability will be shown}

\item{...}{Further parameters for the \code{plot}-function in package \code{igraph}}
}
\description{
Plots a \code{MarkovChain} object
}
\section{Methods}{
 \describe{
\item{list("signature(x = \"MarkovChain\", order = \"numeric\", digits = \"numeric\")")}{ Plots the transition matrix with order \code{order} of a \code{MarkovChain} object as graph. } }
}

\author{
Michael Scholz \email{michael.scholz@th-deg.de}
}
\keyword{methods}
