% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern.r
\docType{methods}
\name{+,Pattern,Pattern-method}
\alias{+,Pattern,Pattern-method}
\title{Concatenates two \code{Pattern} objects}
\usage{
\S4method{+}{Pattern,Pattern}(e1, e2)
}
\arguments{
\item{e1}{First pattern}

\item{e2}{Second pattern}
}
\description{
Concatenates two \code{Pattern} objects
}
\section{Methods}{
 \describe{
\item{list("signature(e1 = \"Pattern\", e2 = \"Pattern\")")}{Concatenates two \code{Pattern} objects.}
}
}

\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}
