% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{states-methods}
\alias{states}
\alias{states,MarkovChain-method}
\alias{states-methods}
\title{Returns all states}
\usage{
states(object)
}
\arguments{
  \item{object}{An instance of the
  \code{MarkovChain}-class}
}
\description{
Returns all states
}
\section{Methods}{
  \describe{

  \item{list("signature(object = \"MarkovChain\")")}{
  Returns the name of all states of a \code{MarkovChain}
  object. } }
}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}

