% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChain.r
\docType{methods}
\name{transientStates}
\alias{transientStates}
\alias{transientStates,MarkovChain-method}
\title{Returns All Transient States}
\usage{
transientStates(object)
}
\arguments{
\item{object}{An instance of the \code{MarkovChain}-class}
}
\description{
Returns All Transient States
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"MarkovChain\")")}{ Returns the names of all states that have a non-zero
probability that a user will never return to them (i.e. that are transient). } }
}

\author{
Michael Scholz \email{michael.scholz@th-deg.de}
}
\keyword{methods}
