\name{transientStates-methods}
\docType{methods}
\alias{transientStates}
\alias{transientStates,MarkovChain-method}
\title{Returns all transient states}
\description{
Returns the names of all states that have a non-zero probability that a user will never return to them (i.e. that are transient).
}
\section{Methods}{
\describe{

\item{\code{signature(object = "MarkovChain")}}{
%%  ~~describe this method here~~
}
}}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}
