% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{absorbingStates-methods}
\alias{absorbingStates}
\alias{absorbingStates,MarkovChain-method}
\alias{absorbingStates-methods}
\title{Returns all absorbing states}
\usage{
absorbingStates(object)
}
\arguments{
  \item{object}{An instance of the
  \code{MarkovChain}-class}
}
\description{
Returns all absorbing states
}
\section{Methods}{
  \describe{

  \item{list("signature(object = \"MarkovChain\")")}{
  Returns the names of all states that never have a
  successor in a clickstream (i.e. that are absorbing).} }
}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}

