% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clickstream.r
\name{print.Clickstreams}
\alias{print.Clickstreams}
\title{Prints a Clickstreams Object}
\usage{
\method{print}{Clickstreams}(x, ...)
}
\arguments{
\item{x}{A list of clickstreams.}

\item{...}{Ignored parameters.}
}
\description{
Prints a \code{Clickstreams} object
}
\examples{

clickstreams <- c("User1,h,c,c,p,c,h,c,p,p,c,p,p,o",
               "User2,i,c,i,c,c,c,d",
               "User3,h,i,c,i,c,p,c,c,p,c,c,i,d",
               "User4,c,c,p,c,d",
               "User5,h,c,c,p,p,c,p,p,p,i,p,o",
               "User6,i,h,c,c,p,p,c,p,c,d")
cls <- as.clickstreams(clickstreams, header = TRUE)
print(cls)

}
\seealso{
\code{\link{readClickstreams}}, \code{\link{randomClickstreams}}
}
\author{
Michael Scholz \email{michael.scholz@th-deg.de}
}
