% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Pattern.r
\docType{methods}
\name{show-methods}
\alias{show,Pattern-method}
\alias{show-methods}
\title{Shows a \code{Pattern} Object}
\usage{
\S4method{show}{Pattern}(object)
}
\arguments{
  \item{object}{An instance of the \code{Pattern}-class}
}
\description{
Shows a \code{Pattern} Object
}
\section{Methods}{
  \describe{ \item{list("signature(object = \"MarkovChain
  \")")}{ Shows a \code{Pattern} object. } }
}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}

