% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChain.r
\docType{methods}
\name{states}
\alias{states}
\alias{states,MarkovChain-method}
\title{Returns All States}
\usage{
states(object)
}
\arguments{
\item{object}{An instance of the \code{MarkovChain}-class}
}
\description{
Returns All States
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"MarkovChain\")")}{ Returns the name of all states of a \code{MarkovChain} object. } }
}

\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}
