\name{states-methods}
\docType{methods}
\alias{states}
\alias{states,MarkovChain-method}
\title{Returns all states}
\description{
Returns the name of all states of a \code{MarkovChain} object.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "MarkovChain")}}{
%%  ~~describe this method here~~
}
}}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}
