% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{summary-methods}
\alias{summary,MarkovChain-method}
\alias{summary-methods}
\title{Prints a summary of a MarkovChain object}
\usage{
\S4method{summary}{MarkovChain}(object)
}
\arguments{
  \item{object}{An instance of the
  \code{MarkovChain}-class}
}
\value{
Returns a \code{MarkovChainSummary} object.

\item{list("desc")}{A short description of the
\code{MarkovChain} object.} \item{list("observations")}{The
number of observations from which the \code{MarkovChain}
object has been fitted.} \item{list("k")}{The number of
estimation parameters.} \item{list("logLikelihood")}{The
LogLikelihood of the \code{MarkovChain} object.}
\item{list("aic")}{Akaike's Information Criterion for the
\code{MarkovChain} object} \item{list("bic")}{Bayesian
Information Criterion for the \code{MarkovChain} object}
}
\description{
Prints a summary of a MarkovChain object
}
\section{Methods}{
  \describe{

  \item{list("signature(object = \"MarkovChain\")")}{
  Generates a summary for a given \code{MarkovChain} object
  } }
}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}

