% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Clickstream.r
\name{summary.Clickstreams}
\alias{summary.Clickstreams}
\title{Prints a Summary of a Clickstreams Object}
\usage{
\method{summary}{Clickstreams}(object, ...)
}
\arguments{
  \item{object}{A \code{Clickstreams} object (see
  \code{\link{readClickstreams}}).} \item{...}{Ignored
  parameters.}
}
\description{
Prints a summary of a \code{Clickstreams} object.
}
\examples{
clickstreams <- c("User1,h,c,c,p,c,h,c,p,p,c,p,p,o",
               "User2,i,c,i,c,c,c,d",
               "User3,h,i,c,i,c,p,c,c,p,c,c,i,d",
               "User4,c,c,p,c,d",
               "User5,h,c,c,p,p,c,p,p,p,i,p,o",
               "User6,i,h,c,c,p,p,c,p,c,d")
csf <- tempfile()
writeLines(clickstreams, csf)
cls <- readClickstreams(csf, header = TRUE)
summary(cls)
}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\seealso{
\code{\link{readClickstreams}},
\code{\link{randomClickstreams}}
}

