% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{transientStates-methods}
\alias{transientStates}
\alias{transientStates,MarkovChain-method}
\alias{transientStates-methods}
\title{Returns all transient states}
\usage{
transientStates(object)
}
\arguments{
  \item{object}{An instance of the
  \code{MarkovChain}-class}
}
\description{
Returns all transient states
}
\section{Methods}{
  \describe{

  \item{list("signature(object = \"MarkovChain\")")}{
  Returns the names of all states that have a non-zero
  probability that a user will never return to them (i.e.
  that are transient). } }
}
\author{
Michael Scholz \email{michael.scholz@uni-passau.de}
}
\keyword{methods}

